Bump JDK baseline to 17
