+++ .github/actions/setup-test-jdk/action.yml
@@ -11,7 +11,7 @@ runs:
     - uses: actions/setup-java@c5195efecf7bdfc987ee8bae7a71cb8b11521c00 # v4.7.1
       with:
         distribution: ${{ inputs.distribution }}
-        java-version: 8
+        java-version: 17
         check-latest: true
     - shell: bash
-      run: echo "JDK8=$JAVA_HOME" >> $GITHUB_ENV
+      run: echo "JDK17=$JAVA_HOME" >> $GITHUB_ENV
