+++ gradle.properties
@@ -14,7 +14,7 @@ org.gradle.jvmargs=-Xmx1g -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryEr
 org.gradle.caching=true
 org.gradle.parallel=true
 org.gradle.configuration-cache.parallel=true
-org.gradle.java.installations.fromEnv=GRAALVM_HOME,JDK8,JDK18,JDK19,JDK20,JDK21,JDK22,JDK23,JDK24
+org.gradle.java.installations.fromEnv=GRAALVM_HOME,JDK17,JDK21,JDK24,JDK25
 org.gradle.kotlin.dsl.allWarningsAsErrors=true
 
 # Test Distribution
