+++ gradle/plugins/common/src/main/kotlin/JavaLibraryExtension.kt
@@ -1,7 +1,7 @@
 import org.gradle.api.JavaVersion
 
 open class JavaLibraryExtension {
-    var mainJavaVersion: JavaVersion = JavaVersion.VERSION_1_8
+    var mainJavaVersion: JavaVersion = JavaVersion.VERSION_17
     var testJavaVersion: JavaVersion = JavaVersion.VERSION_21
     var configureRelease: Boolean = true
 }
