+++ junit-platform-commons/src/main/java/org/junit/platform/commons/util/ModuleUtils.java
@@ -10,18 +10,34 @@
 
 package org.junit.platform.commons.util;
 
-import static java.util.Collections.emptyList;
-import static java.util.Collections.emptySet;
+import static java.util.function.Predicate.isEqual;
+import static java.util.stream.Collectors.toCollection;
+import static java.util.stream.Collectors.toSet;
 import static org.apiguardian.api.API.Status.INTERNAL;
 
+import java.io.IOException;
+import java.lang.module.Configuration;
+import java.lang.module.ModuleFinder;
+import java.lang.module.ModuleReader;
+import java.lang.module.ModuleReference;
+import java.lang.module.ResolvedModule;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.apiguardian.api.API;
+import org.junit.platform.commons.JUnitException;
 import org.junit.platform.commons.logging.Logger;
 import org.junit.platform.commons.logging.LoggerFactory;
+import org.junit.platform.commons.support.DefaultResource;
 import org.junit.platform.commons.support.Resource;
 import org.junit.platform.commons.support.scanning.ClassFilter;
 
@@ -49,42 +65,40 @@ public class ModuleUtils {
 	 * potentially empty
 	 */
 	public static Set<String> findAllNonSystemBootModuleNames() {
-		logger.config(() -> "Basic version of findAllNonSystemBootModuleNames() always returns an empty set!");
-		return emptySet();
+		// @formatter:off
+		Set<String> systemModules = ModuleFinder.ofSystem().findAll().stream()
+				.map(reference -> reference.descriptor().name())
+				.collect(toSet());
+		return streamResolvedModules(name -> !systemModules.contains(name))
+				.map(ResolvedModule::name)
+				.collect(toCollection(LinkedHashSet::new));
+		// @formatter:on
 	}
 
 	/**
-	 * Determine if the current Java runtime supports the Java Platform Module System.
+	 * Java 9+ runtime supports the Java Platform Module System.
 	 *
-	 * @return {@code true} if the Java Platform Module System is available,
-	 * otherwise {@code false}
+	 * @return {@code true}
 	 */
 	public static boolean isJavaPlatformModuleSystemAvailable() {
-		return false;
+		return true;
 	}
 
-	/**
-	 * Return the name of the module that the class or interface is a member of.
-	 *
-	 * @param type class or interface to analyze
-	 * @return the module name; never {@code null} but potentially empty
-	 */
 	public static Optional<String> getModuleName(Class<?> type) {
-		return Optional.empty();
+		Preconditions.notNull(type, "Class type must not be null");
+
+		return Optional.ofNullable(type.getModule().getName());
 	}
 
-	/**
-	 * Return the raw version of the module that the class or interface is a member of.
-	 *
-	 * @param type class or interface to analyze
-	 * @return the raw module version; never {@code null} but potentially empty
-	 */
 	public static Optional<String> getModuleVersion(Class<?> type) {
-		return Optional.empty();
+		Preconditions.notNull(type, "Class type must not be null");
+
+		Module module = type.getModule();
+		return module.isNamed() ? module.getDescriptor().rawVersion() : Optional.empty();
 	}
 
 	/**
-	 * Find all classes for the given module name.
+	 * Find all {@linkplain Class classes} for the given module name.
 	 *
 	 * @param moduleName the name of the module to scan; never {@code null} or
 	 * <em>empty</em>
@@ -96,12 +110,17 @@ public class ModuleUtils {
 		Preconditions.notBlank(moduleName, "Module name must not be null or empty");
 		Preconditions.notNull(filter, "Class filter must not be null");
 
-		logger.config(() -> "Basic version of findAllClassesInModule() always returns an empty list!");
-		return emptyList();
+		logger.debug(() -> "Looking for classes in module: " + moduleName);
+		// @formatter:off
+		Set<ModuleReference> moduleReferences = streamResolvedModules(isEqual(moduleName))
+				.map(ResolvedModule::reference)
+				.collect(toSet());
+		// @formatter:on
+		return scan(moduleReferences, filter, ModuleUtils.class.getClassLoader());
 	}
 
 	/**
-	 * Find all resources for the given module name.
+	 * Find all {@linkplain Resource resources} for the given module name.
 	 *
 	 * @param moduleName the name of the module to scan; never {@code null} or
 	 * <em>empty</em>
@@ -115,8 +134,178 @@ public class ModuleUtils {
 		Preconditions.notBlank(moduleName, "Module name must not be null or empty");
 		Preconditions.notNull(filter, "Resource filter must not be null");
 
-		logger.config(() -> "Basic version of findAllResourcesInModule() always returns an empty list!");
-		return emptyList();
+		logger.debug(() -> "Looking for classes in module: " + moduleName);
+		// @formatter:off
+		Set<ModuleReference> moduleReferences = streamResolvedModules(isEqual(moduleName))
+				.map(ResolvedModule::reference)
+				.collect(toSet());
+		// @formatter:on
+		return scan(moduleReferences, filter, ModuleUtils.class.getClassLoader());
+	}
+
+	/**
+	 * Stream resolved modules from current (or boot) module layer.
+	 */
+	private static Stream<ResolvedModule> streamResolvedModules(Predicate<String> moduleNamePredicate) {
+		Module module = ModuleUtils.class.getModule();
+		ModuleLayer layer = module.getLayer();
+		if (layer == null) {
+			logger.config(() -> ModuleUtils.class + " is a member of " + module
+					+ " - using boot layer returned by ModuleLayer.boot() as fall-back.");
+			layer = ModuleLayer.boot();
+		}
+		return streamResolvedModules(moduleNamePredicate, layer);
+	}
+
+	/**
+	 * Stream resolved modules from the supplied layer.
+	 */
+	private static Stream<ResolvedModule> streamResolvedModules(Predicate<String> moduleNamePredicate,
+			ModuleLayer layer) {
+		logger.debug(() -> "Streaming modules for layer @" + System.identityHashCode(layer) + ": " + layer);
+		Configuration configuration = layer.configuration();
+		logger.debug(() -> "Module layer configuration: " + configuration);
+		Stream<ResolvedModule> stream = configuration.modules().stream();
+		return stream.filter(module -> moduleNamePredicate.test(module.name()));
+	}
+
+	/**
+	 * Scan for classes using the supplied set of module references, class
+	 * filter, and loader.
+	 */
+	private static List<Class<?>> scan(Set<ModuleReference> references, ClassFilter filter, ClassLoader loader) {
+		logger.debug(() -> "Scanning " + references.size() + " module references: " + references);
+		ModuleReferenceClassScanner scanner = new ModuleReferenceClassScanner(filter, loader);
+		List<Class<?>> classes = new ArrayList<>();
+		for (ModuleReference reference : references) {
+			classes.addAll(scanner.scan(reference));
+		}
+		logger.debug(() -> "Found " + classes.size() + " classes: " + classes);
+		return Collections.unmodifiableList(classes);
+	}
+
+	/**
+	 * Scan for classes using the supplied set of module references, class
+	 * filter, and loader.
+	 */
+	private static List<Resource> scan(Set<ModuleReference> references, Predicate<Resource> filter,
+			ClassLoader loader) {
+		logger.debug(() -> "Scanning " + references.size() + " module references: " + references);
+		ModuleReferenceResourceScanner scanner = new ModuleReferenceResourceScanner(filter, loader);
+		List<Resource> classes = new ArrayList<>();
+		for (ModuleReference reference : references) {
+			classes.addAll(scanner.scan(reference));
+		}
+		logger.debug(() -> "Found " + classes.size() + " classes: " + classes);
+		return Collections.unmodifiableList(classes);
+	}
+
+	/**
+	 * {@link ModuleReference} class scanner.
+	 *
+	 * @since 1.1
+	 */
+	static class ModuleReferenceClassScanner {
+
+		private final ClassFilter classFilter;
+		private final ClassLoader classLoader;
+
+		ModuleReferenceClassScanner(ClassFilter classFilter, ClassLoader classLoader) {
+			this.classFilter = classFilter;
+			this.classLoader = classLoader;
+		}
+
+		/**
+		 * Scan module reference for classes that potentially contain testable methods.
+		 */
+		List<Class<?>> scan(ModuleReference reference) {
+			try (ModuleReader reader = reference.open()) {
+				try (Stream<String> names = reader.list()) {
+					// @formatter:off
+					return names.filter(name -> name.endsWith(".class"))
+							.map(this::className)
+							.filter(name -> !name.equals("module-info"))
+							.filter(classFilter::match)
+							.map(this::loadClassUnchecked)
+							.filter(classFilter::match)
+							.collect(Collectors.toList());
+					// @formatter:on
+				}
+			}
+			catch (IOException e) {
+				throw new JUnitException("Failed to read contents of " + reference + ".", e);
+			}
+		}
+
+		/**
+		 * Convert resource name to binary class name.
+		 */
+		private String className(String resourceName) {
+			resourceName = resourceName.substring(0, resourceName.length() - 6); // 6 = ".class".length()
+			resourceName = resourceName.replace('/', '.');
+			return resourceName;
+		}
+
+		/**
+		 * Load class by its binary name.
+		 *
+		 * @see ClassLoader#loadClass(String)
+		 */
+		private Class<?> loadClassUnchecked(String binaryName) {
+			try {
+				return classLoader.loadClass(binaryName);
+			}
+			catch (ClassNotFoundException e) {
+				throw new JUnitException("Failed to load class with name '" + binaryName + "'.", e);
+			}
+		}
+
+	}
+
+	/**
+	 * {@link ModuleReference} resource class scanner.
+	 *
+	 * @since 1.11
+	 */
+	static class ModuleReferenceResourceScanner {
+
+		private final Predicate<Resource> resourceFilter;
+		private final ClassLoader classLoader;
+
+		ModuleReferenceResourceScanner(Predicate<Resource> resourceFilter, ClassLoader classLoader) {
+			this.resourceFilter = resourceFilter;
+			this.classLoader = classLoader;
+		}
+
+		/**
+		 * Scan module reference for resources that potentially contain testable resources.
+		 */
+		List<Resource> scan(ModuleReference reference) {
+			try (ModuleReader reader = reference.open()) {
+				try (Stream<String> names = reader.list()) {
+					// @formatter:off
+					return names.filter(name -> !name.endsWith(".class"))
+							.map(this::loadResourceUnchecked)
+							.filter(resourceFilter)
+							.collect(Collectors.toList());
+					// @formatter:on
+				}
+			}
+			catch (IOException e) {
+				throw new JUnitException("Failed to read contents of " + reference + ".", e);
+			}
+		}
+
+		private Resource loadResourceUnchecked(String binaryName) {
+			try {
+				URI uri = classLoader.getResource(binaryName).toURI();
+				return new DefaultResource(binaryName, uri);
+			}
+			catch (URISyntaxException e) {
+				throw new JUnitException("Failed to load resource with name '" + binaryName + "'.", e);
+			}
+		}
+
 	}
 
 }
