--- junit-platform-commons/src/main/java9/org/junit/platform/commons/util/PackageNameUtils.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2015-2025 the original author or authors.
- *
- * All rights reserved. This program and the accompanying materials are
- * made available under the terms of the Eclipse Public License v2.0 which
- * accompanies this distribution and is available at
- *
- * https://www.eclipse.org/legal/epl-v20.html
- */
-
-package org.junit.platform.commons.util;
-
-/**
- * Collection of utilities for working with package names.
- *
- * <h2>DISCLAIMER</h2>
- *
- * <p>These utilities are intended solely for usage within the JUnit framework
- * itself. <strong>Any usage by external parties is not supported.</strong>
- * Use at your own risk!
- *
- * @since 1.11.3
- */
-class PackageNameUtils {
-
-	static String getPackageName(Class<?> clazz) {
-		return clazz.getPackageName();
-	}
-
-}
