+++ junit-platform-console-standalone/junit-platform-console-standalone.gradle.kts
@@ -86,12 +86,4 @@ tasks {
 			))
 		}
 	}
-
-	// This jar contains some Java 9 code
-	// (org.junit.platform.console.ConsoleLauncherToolProvider which implements
-	// java.util.spi.ToolProvider which is @since 9).
-	// So in order to resolve this, it can only run on Java 9
-	osgiProperties {
-		property("-runee", "JavaSE-9")
-	}
 }
