+++ junit-platform-console/junit-platform-console.gradle.kts
@@ -3,7 +3,6 @@ import junitbuild.java.UpdateJarAction
 plugins {
 	id("junitbuild.java-library-conventions")
 	id("junitbuild.shadow-conventions")
-	id("junitbuild.java-multi-release-sources")
 }
 
 description = "JUnit Platform Console"
@@ -33,22 +32,17 @@ tasks {
 		))
 	}
 	shadowJar {
-		val release17ClassesDir = sourceSets.mainRelease17.get().output.classesDirs.singleFile
-		inputs.dir(release17ClassesDir).withPathSensitivity(PathSensitivity.RELATIVE)
 		exclude("META-INF/versions/9/module-info.class")
 		relocate("picocli", "org.junit.platform.console.shadow.picocli")
 		from(projectDir) {
 			include("LICENSE-picocli.md")
 			into("META-INF")
 		}
-		from(sourceSets.mainRelease9.get().output.classesDirs)
 		doLast(objects.newInstance(UpdateJarAction::class).apply {
 			javaLauncher = project.javaToolchains.launcherFor(java.toolchain)
 			args.addAll(
 				"--file", archiveFile.get().asFile.absolutePath,
 				"--main-class", "org.junit.platform.console.ConsoleLauncher",
-				"--release", "17",
-				"-C", release17ClassesDir.absolutePath, "."
 			)
 		})
 	}
@@ -59,20 +53,5 @@ tasks {
 		manifest {
 			attributes("Main-Class" to "org.junit.platform.console.ConsoleLauncher")
 		}
-	}
-
-	// This jar contains some Java 9 code
-	// (org.junit.platform.console.ConsoleLauncherToolProvider which implements
-	// java.util.spi.ToolProvider which is @since 9).
-	// So in order to resolve this, it can only run on Java 9
-	osgiProperties {
-		property("-runee", "JavaSE-9")
-	}
-}
-
-eclipse {
-	classpath {
-		sourceSets -= project.sourceSets.mainRelease9.get()
-		sourceSets -= project.sourceSets.mainRelease17.get()
 	}
 }
