+++ junit-platform-console/src/main/java/org/junit/platform/console/options/ConsoleUtils.java
@@ -12,6 +12,7 @@ package org.junit.platform.console.options;
 
 import static org.apiguardian.api.API.Status.INTERNAL;
 
+import java.io.Console;
 import java.nio.charset.Charset;
 
 import org.apiguardian.api.API;
@@ -35,7 +36,8 @@ public class ConsoleUtils {
 	 * {@return the charset of the console}
 	 */
 	public static Charset charset() {
-		return Charset.defaultCharset();
+		Console console = System.console();
+		return console != null ? console.charset() : Charset.defaultCharset();
 	}
 
 }
