+++ junit-platform-jfr/junit-platform-jfr.gradle.kts
@@ -20,17 +20,7 @@ dependencies {
 	osgiVerification(projects.junitPlatformLauncher)
 }
 
-javaLibrary {
-	// --release 8 does not support jdk.jfr even though it was backported
-	configureRelease = false
-}
-
 tasks {
-	compileJava {
-		javaCompiler = project.javaToolchains.compilerFor {
-			languageVersion = JavaLanguageVersion.of(8)
-		}
-	}
 	compileModule {
 		options.release = 11
 	}
