+++ platform-tooling-support-tests/platform-tooling-support-tests.gradle.kts
@@ -211,12 +211,12 @@ val test by testing.suites.getting(JvmTestSuite::class) {
 
 				develocity {
 					testDistribution {
-						requirements.add("jdk=8")
+						requirements.add("jdk=17")
 						this as TestDistributionConfigurationInternal
 						preferredMaxDuration = Duration.ofMillis(500)
 					}
 				}
-				jvmArgumentProviders += JavaHomeDir(project, 8, develocity.testDistribution.enabled)
+				jvmArgumentProviders += JavaHomeDir(project, 17, develocity.testDistribution.enabled)
 				jvmArgumentProviders += JavaHomeDir(project, 17, develocity.testDistribution.enabled)
 
 				val gradleJavaVersion = JavaVersion.current().majorVersion.toInt()
