+++ platform-tooling-support-tests/projects/gradle-kotlin-extensions/gradle.properties
@@ -1,1 +1,1 @@
-org.gradle.java.installations.fromEnv=JDK8
+org.gradle.java.installations.fromEnv=JDK17
