+++ platform-tooling-support-tests/projects/gradle-missing-engine/build.gradle.kts
@@ -18,7 +18,7 @@ dependencies {
 
 java {
 	toolchain {
-		languageVersion = JavaLanguageVersion.of(8)
+		languageVersion = JavaLanguageVersion.of(17)
 	}
 }
