+++ platform-tooling-support-tests/projects/jupiter-starter/gradle.properties
@@ -1,1 +1,1 @@
-org.gradle.java.installations.fromEnv=JDK8,JDK17
+org.gradle.java.installations.fromEnv=JDK17
