+++ platform-tooling-support-tests/projects/jupiter-starter/pom.xml
@@ -9,8 +9,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
+        <maven.compiler.release>17</maven.compiler.release>
         <junit.platform.commons.version>${junit.platform.version}</junit.platform.commons.version>
     </properties>
