+++ platform-tooling-support-tests/projects/vintage/build.gradle.kts
@@ -26,7 +26,7 @@ dependencies {
 
 java {
 	toolchain {
-		languageVersion = JavaLanguageVersion.of(8)
+		languageVersion = JavaLanguageVersion.of(17)
 	}
 }
