+++ platform-tooling-support-tests/src/test/java/platform/tooling/support/tests/GradleMissingEngineTests.java
@@ -37,7 +37,7 @@ class GradleMissingEngineTests {
 				.workingDir(copyToWorkspace(Projects.GRADLE_MISSING_ENGINE, workspace)) //
 				.addArguments("-Dmaven.repo=" + MavenRepo.dir()) //
 				.addArguments("build", "--no-daemon", "--stacktrace", "--no-build-cache", "--warning-mode=fail") //
-				.putEnvironment("JDK8", Helper.getJavaHome(8).orElseThrow(TestAbortedException::new).toString()) //
+				.putEnvironment("JDK17", Helper.getJavaHome(17).orElseThrow(TestAbortedException::new).toString()) //
 				.redirectOutput(outputFiles).startAndWait();
 
 		assertEquals(1, result.exitCode());
