+++ platform-tooling-support-tests/src/test/java/platform/tooling/support/tests/ManifestTests.java
@@ -56,10 +56,8 @@ class ManifestTests {
 			assertValue(attributes, "Bundle-SymbolicName", module);
 			assertValue(attributes, "Bundle-Version",
 				MavenVersion.parseMavenString(version).getOSGiVersion().toString());
-			switch (module) {
-				case "junit-platform-commons" -> assertValue(attributes, "Multi-Release", "true");
-				case "junit-platform-console" -> assertValue(attributes, "Main-Class",
-					"org.junit.platform.console.ConsoleLauncher");
+			if (module.equals("junit-platform-console")) {
+				assertValue(attributes, "Main-Class", "org.junit.platform.console.ConsoleLauncher");
 			}
 			var domain = Domain.domain(manifest);
 			domain.getExportPackage().forEach((pkg, attrs) -> {
